requirements.txt
# 최소 필수 패키지만

# 웹 프레임워크
fastapi==0.109.0
uvicorn[standard]==0.27.0

# 슬랙
slack-bolt==1.18.1

# Google Sheets
gspread==5.12.0
google-auth==2.25.2

# Gemini
google-generativeai==0.3.2

# 크롤링
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# 기타
python-dotenv==1.0.0
